+++ docs/source/conf.py
@@ -656,7 +656,7 @@ def create_doctest_figure(app, obj, name, lines):
         num_lines = part.count('\n')
 
         doctests = list(xdoctest.core.parse_docstr_examples(
-            part, modpath=modpath, callname=name))
+            part, modpath=modpath, callname=name, style='google'))
         # print(doctests)
 
         # doctests = list(xdoctest.core.parse_docstr_examples(